+++ sdk/keyvault/azure-security-keyvault-administration/src/main/java/com/azure/security/keyvault/administration/KeyVaultAccessControlClientBuilder.java
@@ -95,7 +95,7 @@ public final class KeyVaultAccessControlClientBuilder implements
     private Configuration configuration;
     private ClientOptions clientOptions;
     private KeyVaultAdministrationServiceVersion serviceVersion;
-    private boolean verifyChallengeResource = true;
+    private boolean disableChallengeResourceVerification = false;
 
     /**
      * Creates a {@link KeyVaultAccessControlClientBuilder} instance that is able to configure and construct
@@ -184,7 +184,7 @@ public final class KeyVaultAccessControlClientBuilder implements
         // Add retry policy.
         policies.add(ClientBuilderUtil.validateAndGetRetryPolicy(retryPolicy, retryOptions));
 
-        policies.add(new KeyVaultCredentialPolicy(credential, verifyChallengeResource));
+        policies.add(new KeyVaultCredentialPolicy(credential, disableChallengeResourceVerification));
 
         // Add per retry additional policies.
         policies.addAll(perRetryPolicies);
@@ -444,12 +444,13 @@ public final class KeyVaultAccessControlClientBuilder implements
     /** Gets or sets whether to verify the authentication challenge resource matches the Key Vault or Managed HSM
      * domain. The default is set to {@code true}.
      *
-     * @param verifyChallengeResource A flag indicating if the authentication challenge resource must be verified.
+     * @param disableChallengeResourceVerification A flag indicating if the authentication challenge resource must be
+     * verified.
      *
      * @return The updated {@link KeyVaultAccessControlClientBuilder} object.
      */
-    public KeyVaultAccessControlClientBuilder verifyChallengeResource(boolean verifyChallengeResource) {
-        this.verifyChallengeResource = verifyChallengeResource;
+    public KeyVaultAccessControlClientBuilder disableChallengeResourceVerification(boolean disableChallengeResourceVerification) {
+        this.disableChallengeResourceVerification = disableChallengeResourceVerification;
 
         return this;
     }