+++ sdk/keyvault/azure-security-keyvault-administration/src/main/java/com/azure/security/keyvault/administration/KeyVaultBackupClientBuilder.java
@@ -94,7 +94,7 @@ public final class KeyVaultBackupClientBuilder implements
     private Configuration configuration;
     private ClientOptions clientOptions;
     private KeyVaultAdministrationServiceVersion serviceVersion;
-    private boolean verifyChallengeResource = true;
+    private boolean disableChallengeResourceVerification = false;
 
     /**
      * Creates a {@link KeyVaultBackupClientBuilder} instance that is able to configure and construct instances of
@@ -183,7 +183,7 @@ public final class KeyVaultBackupClientBuilder implements
         // Add retry policy.
         policies.add(ClientBuilderUtil.validateAndGetRetryPolicy(retryPolicy, retryOptions));
 
-        policies.add(new KeyVaultCredentialPolicy(credential, verifyChallengeResource));
+        policies.add(new KeyVaultCredentialPolicy(credential, disableChallengeResourceVerification));
 
         // Add per retry additional policies.
         policies.addAll(perRetryPolicies);
@@ -443,12 +443,13 @@ public final class KeyVaultBackupClientBuilder implements
     /** Gets or sets whether to verify the authentication challenge resource matches the Key Vault or Managed HSM
      * domain. The default is set to {@code true}.
      *
-     * @param verifyChallengeResource A flag indicating if the authentication challenge resource must be verified.
+     * @param disableChallengeResourceVerification A flag indicating if the authentication challenge resource must be
+     * verified.
      *
      * @return The updated {@link KeyVaultBackupClientBuilder} object.
      */
-    public KeyVaultBackupClientBuilder verifyChallengeResource(boolean verifyChallengeResource) {
-        this.verifyChallengeResource = verifyChallengeResource;
+    public KeyVaultBackupClientBuilder disableChallengeResourceVerification(boolean disableChallengeResourceVerification) {
+        this.disableChallengeResourceVerification = disableChallengeResourceVerification;
 
         return this;
     }