+++ sdk/keyvault/azure-security-keyvault-administration/src/main/java/com/azure/security/keyvault/administration/implementation/KeyVaultCredentialPolicy.java
@@ -41,17 +41,17 @@ public class KeyVaultCredentialPolicy extends BearerTokenAuthenticationPolicy {
     private static final String WWW_AUTHENTICATE = "WWW-Authenticate";
     private static final ConcurrentMap<String, ChallengeParameters> CHALLENGE_CACHE = new ConcurrentHashMap<>();
     private ChallengeParameters challenge;
-    private final boolean verifyChallengeResource;
+    private final boolean disableChallengeResourceVerification;
 
     /**
      * Creates a {@link KeyVaultCredentialPolicy}.
      *
      * @param credential The token credential to authenticate the request.
      */
-    public KeyVaultCredentialPolicy(TokenCredential credential, boolean verifyChallengeResource) {
+    public KeyVaultCredentialPolicy(TokenCredential credential, boolean disableChallengeResourceVerification) {
         super(credential);
 
-        this.verifyChallengeResource = verifyChallengeResource;
+        this.disableChallengeResourceVerification = disableChallengeResourceVerification;
     }
 
     /**
@@ -167,7 +167,7 @@ public class KeyVaultCredentialPolicy extends BearerTokenAuthenticationPolicy {
                     return Mono.just(false);
                 }
             } else {
-                if (verifyChallengeResource) {
+                if (!disableChallengeResourceVerification) {
                     if (!isChallengeResourceValid(request, scope)) {
                         throw LOGGER.logExceptionAsError(
                             new RuntimeException(String.format(
@@ -275,7 +275,7 @@ public class KeyVaultCredentialPolicy extends BearerTokenAuthenticationPolicy {
                 return false;
             }
         } else {
-            if (verifyChallengeResource) {
+            if (!disableChallengeResourceVerification) {
                 if (!isChallengeResourceValid(request, scope)) {
                     throw LOGGER.logExceptionAsError(
                         new RuntimeException(String.format(