+++ sdk/keyvault/azure-security-keyvault-certificates/src/main/java/com/azure/security/keyvault/certificates/CertificateClientBuilder.java
@@ -132,7 +132,7 @@ public final class CertificateClientBuilder implements
     private Configuration configuration;
     private CertificateServiceVersion version;
     private ClientOptions clientOptions;
-    private boolean verifyChallengeResource = true;
+    private boolean disableChallengeResourceVerification = false;
 
     /**
      * The constructor with defaults.
@@ -231,7 +231,7 @@ public final class CertificateClientBuilder implements
         // Add retry policy.
         policies.add(ClientBuilderUtil.validateAndGetRetryPolicy(retryPolicy, retryOptions));
 
-        policies.add(new KeyVaultCredentialPolicy(credential, verifyChallengeResource));
+        policies.add(new KeyVaultCredentialPolicy(credential, disableChallengeResourceVerification));
 
         // Add per retry additional policies.
         policies.addAll(perRetryPolicies);
@@ -491,12 +491,13 @@ public final class CertificateClientBuilder implements
     /** Gets or sets whether to verify the authentication challenge resource matches the Key Vault or Managed HSM
      * domain. The default is set to {@code true}.
      *
-     * @param verifyChallengeResource A flag indicating if the authentication challenge resource must be verified.
+     * @param disableChallengeResourceVerification A flag indicating if the authentication challenge resource must be
+     * verified.
      *
      * @return The updated {@link CertificateClientBuilder} object.
      */
-    public CertificateClientBuilder verifyChallengeResource(boolean verifyChallengeResource) {
-        this.verifyChallengeResource = verifyChallengeResource;
+    public CertificateClientBuilder disableChallengeResourceVerification(boolean disableChallengeResourceVerification) {
+        this.disableChallengeResourceVerification = disableChallengeResourceVerification;
 
         return this;
     }