+++ sdk/keyvault/azure-security-keyvault-keys/src/main/java/com/azure/security/keyvault/keys/KeyClientBuilder.java
@@ -128,7 +128,7 @@ public final class KeyClientBuilder implements
     private Configuration configuration;
     private KeyServiceVersion version;
     private ClientOptions clientOptions;
-    private boolean verifyChallengeResource = true;
+    private boolean disableChallengeResourceVerification = false;
 
     /**
      * The constructor with defaults.
@@ -230,7 +230,7 @@ public final class KeyClientBuilder implements
         // Add retry policy.
         policies.add(ClientBuilderUtil.validateAndGetRetryPolicy(retryPolicy, retryOptions));
 
-        policies.add(new KeyVaultCredentialPolicy(credential, verifyChallengeResource));
+        policies.add(new KeyVaultCredentialPolicy(credential, disableChallengeResourceVerification));
         // Add per retry additional policies.
         policies.addAll(perRetryPolicies);
 
@@ -491,12 +491,13 @@ public final class KeyClientBuilder implements
     /** Gets or sets whether to verify the authentication challenge resource matches the Key Vault or Managed HSM
      * domain. The default is set to {@code true}.
      *
-     * @param verifyChallengeResource A flag indicating if the authentication challenge resource must be verified.
+     * @param disableChallengeResourceVerification A flag indicating if the authentication challenge resource must be
+     * verified.
      *
      * @return The updated {@link KeyClientBuilder} object.
      */
-    public KeyClientBuilder verifyChallengeResource(boolean verifyChallengeResource) {
-        this.verifyChallengeResource = verifyChallengeResource;
+    public KeyClientBuilder disableChallengeResourceVerification(boolean disableChallengeResourceVerification) {
+        this.disableChallengeResourceVerification = disableChallengeResourceVerification;
 
         return this;
     }