+++ sdk/keyvault/azure-security-keyvault-keys/src/main/java/com/azure/security/keyvault/keys/cryptography/KeyEncryptionKeyClientBuilder.java
@@ -391,12 +391,13 @@ public final class KeyEncryptionKeyClientBuilder implements KeyEncryptionKeyReso
     /** Gets or sets whether to verify the authentication challenge resource matches the Key Vault or Managed HSM
      * domain. The default is set to {@code true}.
      *
-     * @param verifyChallengeResource A flag indicating if the authentication challenge resource must be verified.
+     * @param disableChallengeResourceVerification A flag indicating if the authentication challenge resource must be
+     * verified.
      *
      * @return The updated {@link KeyEncryptionKeyClientBuilder} object.
      */
-    public KeyEncryptionKeyClientBuilder verifyChallengeResource(boolean verifyChallengeResource) {
-        builder.verifyChallengeResource(verifyChallengeResource);
+    public KeyEncryptionKeyClientBuilder disableChallengeResourceVerification(boolean disableChallengeResourceVerification) {
+        builder.disableChallengeResourceVerification(disableChallengeResourceVerification);
 
         return this;
     }