+++ sdk/keyvault/azure-security-keyvault-certificates/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 4.4.0 (2022-09-19)
 
 ### Breaking Changes
-- Made it so that we verify the challenge resource matches the vault domain by default. This should affect few customers who can pass `false` to the `verifyChallengeResource()` method in client builders to disable this functionality. See https://aka.ms/azsdk/blog/vault-uri for more information.
+- Made it so that we verify that the challenge resource matches the vault domain by default. This should affect few customers who can pass `true` to the `disableChallengeResourceVerification()` method in client builders to disable this functionality. See https://aka.ms/azsdk/blog/vault-uri for more information.
 
 ### Other Changes